requirements.txt
absl-py==1.2.0
ale-py==0.7.5
autopep8==1.7.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
cloudpickle==2.1.0
cycler==0.11.0
eclipse-sumo==1.14.1
fonttools==4.36.0
google-auth==2.10.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
gym==0.21.0
gym-notices==0.0.8
idna==3.3
importlib-resources==5.9.0
joblib==1.1.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.23.2
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pygame==2.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.0
six==1.16.0
stable-baselines3==1.6.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
toml==0.10.2
torch==1.12.1
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.2
